python_test/requirements.txt
pytest
playwright
pytest-playwright
python-dotenv
